Fly effect integrated & many bugs fixed.

+++ FeLib/Include/bitmap.h
@@ -161,7 +161,7 @@ class bitmap
   bool IsValidPos(vector2d What) const { return (What.X >= 0 && What.Y >= 0 && What.X < XSize && What.Y < YSize); }
   bool IsValidPos(short X, short Y) const { return (X >= 0 && Y >= 0 && X < XSize && Y < YSize); }
   void CreateSparkle(vector2d, ushort);
-  void CreateFlies(ushort, ulong, ushort);
+  void CreateFlies(ulong, ushort, uchar);
   void CreateLightning();
  protected:
   ushort XSize, YSize;
